refactor(cli): commands to load options
Updates CLI commands to use `loadOptions` function, which attempts to load configuration from the Nuxt project.

If the Nuxt project can't be loaded, or if the nuxt-users configuration is not found, it falls back to environment variables.
This ensures a more flexible configuration loading process.

Removes the dedicated `migrate` command and integrates its functionality into the `create-migrations-table` command, renaming it to `migrate`.